fix err

+++ main_parallel.py
@@ -19,7 +19,7 @@ if __name__ == "__main__":
     parser.add_argument("-ix", "--ix", help = "index for splits", type = int)
     parser.add_argument("-o", "--o", help = "outpath", type = str)
     parser.add_argument("-i", "--i", help = "inpath", type = str)
-    parser.add_argument("-week", "--week", help="week", type=int)
+    parser.add_argument("-week", "--week", help="week", type=float)
     args = parser.parse_args()
     mb = basic_ml()
